fix: move Handshake to own module
